GPII-2297: Added context integration tests for the untrusted flowmanager
