+++ tests/all-tests.js
@@ -49,6 +49,7 @@ var testIncludes = [
     "./PreferencesServerErrorTests.js",
     "./StartupAPITests.js",
     "./UntrustedBrowserChannelTests.js",
+    "./UntrustedContextIntegrationTests.js",
     "./UntrustedDevelopmentTests.js",
     "./PayloadSizeTest.js",
     "./UntrustedUserLogonStateChangeTests.js",
