+++ tests/platform/linux/configs/gpii.tests.acceptance.linux.builtIn.untrusted.config.json
@@ -1,0 +1,24 @@
+{
+    "type": "gpii.tests.acceptance.linux.builtIn.untrusted.config",
+    "options": {
+        "configs": {
+            "cloudBasedConfig": {
+                "configName": "gpii.config.cloudBased.development",
+                "configPath": "%universal/gpii/configs"
+            }
+        },
+        "distributeOptions": {
+            "acceptance.installedSolutionsPath": {
+                "record": "%universal/testData/deviceReporter/acceptanceTests/linux_builtIn.json",
+                "target": "{that localConfig deviceReporter installedSolutionsDataSource}.options.path",
+                "priority": "after:development.installedSolutionsPath"
+            },
+            "acceptance.rawPreferencesDataSource": {
+                "record": "%universal/testData/preferences/acceptanceTests/%userToken.json",
+                "target": "{that cloudBasedConfig rawPreferencesServer rawPreferencesDataSource}.options.path",
+                "priority": "after:development.rawPreferencesDataSource"
+            }
+        }
+    },
+    "mergeConfigs": "%universal/gpii/configs/gpii.config.untrusted.development.all.local.json"
+}
